Refactor: split up podman / iptables stuff
Put these in separate modules, with another to bring them together.
Also separate the config / arg parsing, so the node name doesn't need to
be passed all around the code.
