+++ crates/workerd/src/lib.rs
@@ -5,8 +5,8 @@
 #![allow(clippy::wildcard_imports)]
 
 pub mod cluster_nameserver;
+pub mod containers;
 pub mod limits;
 pub mod pod_claimer;
 pub mod pod_killer;
 pub mod pod_worker;
-pub mod podman;
